backend/app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-python==4.10.0
torch==2.4.1
torchvision==0.19.1
pytesseract==0.3.13
pillow==10.4.0
python-multipart==0.0.9